+++ src/imio/esign/browser/static/esign.js
@@ -1,0 +1,11 @@
+jQuery(function($) {
+    $('a.esign-create-custom-session').prepOverlay({
+        subtype: 'ajax',
+        filter: '#content>*',
+        formselector: '#form',
+        closeselector: '[name="form.buttons.cancel"]',
+        noform: function(el, pbo) {
+            window.location.reload();
+        }
+    });
+});